+++ Platform.PCL/Realm.PCL/RealmConfigurationBasePCL.cs
@@ -28,8 +28,6 @@ namespace Realms
     /// </remarks>
     public class RealmConfigurationBase
     {
-        internal bool Dynamic;
-
         /// <summary>
         /// Gets the filename to be combined with the platform-specific document directory.
         /// </summary>
@@ -41,6 +39,13 @@ namespace Realms
         /// </summary>
         /// <value>The absolute path to the Realm.</value>
         public string DatabasePath { get; }
+
+        /// <summary>
+        /// Gets or sets a value indicating whether the Realm will be open in dynamic mode. If opened in dynamic mode,
+        /// the schema will be read from the file on disk.
+        /// </summary>
+        /// <value><c>true</c> if the Realm will be opened in dynamic mode; <c>false</c> otherwise.</value>
+        public bool IsDynamic { get; set; }
 
         /// <summary>
         /// Gets or sets the list of classes persisted in a Realm opened with this configuration.