+++ CHANGELOG.md
@@ -2,6 +2,7 @@ x.y.z (TBD)
 ------------------
 
 ### Enhancements
+- Added an `IsDynamic` property to `RealmConfigurationBase`, allowing you to open a Realm file and read its schema from disk. ([#1637](https://github.com/realm/realm-dotnet/pull/1637))
 
 ### Bug fixes